refactor: split ETS.py into ETS.py/ETS2.py/_ETS.py
Same treatment as ET/ET2/_ET (previous commit), for the sequence classes:

- _ETS.py: BaseETS only (private/shared - both ETS and ETS2 depend on it,
  so it can't depend on either without a cycle).
- ETS.py: the ETS class.
- ETS2.py: the ETS2 class (imports ET2 directly from ET2.py - safe,
  one-directional, since ET2.py never depends on ETS2.py at load time).

isinstance(self, ETS) inside BaseETS.plot()/teach() (which can't import
the concrete ETS without a cycle) is fixed the same way those methods
already handle Robot/Robot2: a function-local, call-time-deferred import
of ETS from roboticstoolbox.ets.ETS, added alongside the existing local
Robot/Robot2 import.

Mechanical split of a ~2900-line file done by slicing along the existing
class boundaries (BaseETS/ETS/ETS2 were already cleanly delineated) rather
than hand-transcribing; each of the three new files keeps the full
original import block rather than a hand-traced minimal set, trading a
few unused imports per file for confidence against missing something
across this much code.

Verified: full test suite green (610 passed), including test_plot/
test_teach on both ETS and ETS2 specifically (the isinstance fix).
