+++ src/roboticstoolbox/ets/__init__.py
@@ -1,6 +1,7 @@
 from roboticstoolbox.ets.ET import ET
 from roboticstoolbox.ets.ET2 import ET2
-from roboticstoolbox.ets.ETS import ETS, ETS2
+from roboticstoolbox.ets.ETS import ETS
+from roboticstoolbox.ets.ETS2 import ETS2
 
 __all__ = [
     "ET",
