+++ src/roboticstoolbox/robot/Link.py
@@ -14,7 +14,8 @@ from spatialgeometry import Shape, SceneNode, SceneGroup
 from typing import overload
 
 import roboticstoolbox as rtb
-from roboticstoolbox.ets.ETS import ETS, ETS2
+from roboticstoolbox.ets.ETS import ETS
+from roboticstoolbox.ets.ETS2 import ETS2
 from roboticstoolbox.ets.ET import ET
 from roboticstoolbox.ets.ET2 import ET2
 from roboticstoolbox.ets._ET import BaseET
